remove coordinator context
